refactor(extensions): unify 5 per-kind loaders into KindAdapter + ExtensionLoader (W4)
Replace the five separate loader files (user_model_loader, user_vault_loader,
user_driver_loader, user_datastore_loader, user_report_loader) with a single
ExtensionLoader parameterized by a KindAdapter interface. Each kind implements
its own adapter (ModelKindAdapter, VaultKindAdapter, etc.) providing
kind-specific validation, registration, and catalog logic.

Key changes:
- Delete 5 old user_*_loader.ts files (~5000 LOC removed)
- Add KindAdapter interface and 5 implementations (~1400 LOC)
- Add unified ExtensionLoader (~975 LOC)
- Remove legacyStore field from ExtensionRepository
- Migrate 8 construction sites in cli/mod.ts and auto_resolver_adapters.ts
- Add isFullyLoaded() to KindAdapter to distinguish lazy vs fully-loaded entries
- Fix swamp-club#297: I-Repo-1 now skips kind="extension" rows, allowing
  multiple extension files to target the same base model type
