+++ src/cli/commands/doctor_extensions.ts
@@ -130,7 +130,7 @@ export const doctorExtensionsCommand = new Command()
       try {
         rescanRepo.invalidateAll();
       } finally {
-        rescanRepo.legacyStore.close();
+        rescanRepo.close();
       }
     } catch {
       // Best-effort — the loader will bootstrap a fresh catalog if this fails.
