+++ src/libswamp/extensions/reconcile_from_disk_service.ts
@@ -48,11 +48,12 @@ import type {
 import { BUNDLE_LAYOUT_VERSION } from "../../infrastructure/persistence/extension_catalog_store.ts";
 import type { LockfileRepository } from "../../infrastructure/persistence/lockfile_repository.ts";
 import { swampPath } from "../../infrastructure/persistence/paths.ts";
-import { UserModelLoader } from "../../domain/models/user_model_loader.ts";
-import { UserDriverLoader } from "../../domain/drivers/user_driver_loader.ts";
-import { UserVaultLoader } from "../../domain/vaults/user_vault_loader.ts";
-import { UserDatastoreLoader } from "../../domain/datastore/user_datastore_loader.ts";
-import { UserReportLoader } from "../../domain/reports/user_report_loader.ts";
+import { ExtensionLoader } from "../../domain/extensions/extension_loader.ts";
+import { modelKindAdapter } from "../../domain/extensions/model_kind_adapter.ts";
+import { vaultKindAdapter } from "../../domain/extensions/vault_kind_adapter.ts";
+import { driverKindAdapter } from "../../domain/extensions/driver_kind_adapter.ts";
+import { datastoreKindAdapter } from "../../domain/extensions/datastore_kind_adapter.ts";
+import { reportKindAdapter } from "../../domain/extensions/report_kind_adapter.ts";
 import type { DenoRuntime } from "../../domain/runtime/deno_runtime.ts";
 import {
   collectDirsForKind,
@@ -567,35 +568,41 @@ export class ReconcileFromDiskService {
   } {
     switch (kindDir) {
       case "models":
-        return new UserModelLoader(
+        return new ExtensionLoader(
           this.denoRuntime,
+          modelKindAdapter,
           this.repoDir,
           undefined,
           this.repository,
         );
       case "vaults":
-        return new UserVaultLoader(
+        return new ExtensionLoader(
           this.denoRuntime,
+          vaultKindAdapter,
           this.repoDir,
           undefined,
           this.repository,
         );
       case "drivers":
-        return new UserDriverLoader(
+        return new ExtensionLoader(
           this.denoRuntime,
+          driverKindAdapter,
           this.repoDir,
           undefined,
           this.repository,
         );
       case "datastores":
-        return new UserDatastoreLoader(
+        return new ExtensionLoader(
           this.denoRuntime,
+          datastoreKindAdapter,
           this.repoDir,
+          undefined,
           this.repository,
         );
       case "reports":
-        return new UserReportLoader(
+        return new ExtensionLoader(
           this.denoRuntime,
+          reportKindAdapter,
           this.repoDir,
           undefined,
           this.repository,
@@ -604,7 +611,6 @@ export class ReconcileFromDiskService {
   }
 
   private markKindsPopulated(): void {
-    const catalog = this.repository.legacyStore;
     const kinds: ExtensionKind[] = [
       "model",
       "vault",
@@ -612,16 +618,16 @@ export class ReconcileFromDiskService {
       "datastore",
       "report",
     ];
-    catalog.setLayoutVersion(BUNDLE_LAYOUT_VERSION);
+    this.repository.setLayoutVersion(BUNDLE_LAYOUT_VERSION);
     for (const kind of kinds) {
-      catalog.markPopulated(kind);
+      this.repository.markPopulated(kind);
     }
   }
 
   private markAllKindsPopulated(): void {
     this.markKindsPopulated();
     const m = this.localManifestIdentity;
-    this.repository.legacyStore.setManifestIdentity(
+    this.repository.setManifestIdentity(
       m ? `${m.name}@${m.version}` : null,
     );
   }
